Move some non-default skin files to correct namespace
