+++ osu.Game.Rulesets.Osu/Skinning/SnakingSliderBody.cs
@@ -14,7 +14,7 @@ using osu.Game.Rulesets.Osu.Objects;
 using osu.Game.Rulesets.Osu.Objects.Drawables;
 using osuTK;
 
-namespace osu.Game.Rulesets.Osu.Skinning.Default
+namespace osu.Game.Rulesets.Osu.Skinning
 {
     /// <summary>
     /// A <see cref="SliderBody"/> which changes its curve depending on the snaking progress.
